+++ common/src/main/java/org/callimard/printmeacube/common/RegistrationProvider.java
@@ -1,0 +1,5 @@
+package org.callimard.printmeacube.common;
+
+public enum RegistrationProvider {
+    LOCAL, FACEBOOK, GOOGLE, APPLE
+}
