+++ entities/src/main/java/org/callimard/printmeacube/entities/dto/DTOSerializable.java
@@ -1,0 +1,7 @@
+package org.callimard.printmeacube.entities.dto;
+
+public interface DTOSerializable<T> {
+
+    T toDTO();
+
+}
