+++ pom.xml
@@ -3,7 +3,7 @@
     <modelVersion>4.0.0</modelVersion>
     <modules>
         <module>entities</module>
-        <module>authentication</module>
+        <module>user-management</module>
         <module>jwt</module>
         <module>api</module>
         <module>configuration-service</module>
