+++ user-management/pom.xml
@@ -7,10 +7,10 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>authentication</artifactId>
+    <artifactId>user-management</artifactId>
     <packaging>jar</packaging>
 
-    <name>authentication</name>
+    <name>user-management</name>
 
     <properties>
         <maven.compiler.source>19</maven.compiler.source>
