+++ user-management/src/main/java/org/callimard/printmeacube/user_management/authentication/AuthenticationController.java
@@ -1,4 +1,4 @@
-package org.callimard.printmeacube.authentication;
+package org.callimard.printmeacube.user_management.authentication;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import lombok.RequiredArgsConstructor;
