+++ user-management/src/main/java/org/callimard/printmeacube/user_management/authentication/BasicAuthenticationProvider.java
@@ -1,4 +1,4 @@
-package org.callimard.printmeacube.authentication;
+package org.callimard.printmeacube.user_management.authentication;
 
 import com.google.common.collect.Lists;
 import lombok.RequiredArgsConstructor;
