+++ user-management/src/main/resources/application.yml
@@ -1,6 +1,6 @@
 spring:
   application:
-    name: authentication-service
+    name: user-management-service
   cloud:
     config:
       uri: http://localhost:8888
